requirements.txt
Django>=5.0,<6.0
psycopg[binary]>=3.1,<4.0
gunicorn>=22.0,<23.0
django-csp>=3.8,<4.0
django-mfa2>=3.1.0
weasyprint>=62.0,<63.0
pydyf>=0.10,<0.11
python-dotenv>=1.0,<2.0
argon2-cffi>=23.1,<24.0
django-axes>=6.5,<7.0
qrcode>=7.4,<8.0
cryptography>=43.0,<45.0
